Refactor out position sanity checks from controllers
-  Refactor out the Position sanity check code from PositionController
into its own PositionSanityCheckController.
- Update routes to use the new controller.
- Update feature tests to reference the new controller.
