+++ routes/api.php
@@ -210,8 +210,8 @@ Route::group([
 
     Route::get('position/people-by-position', 'PositionController@peopleByPosition');
     Route::get('position/sandman-qualified', 'PositionController@sandmanQualifiedReport');
-    Route::get('position/sanity-checker', 'PositionController@sanityChecker');
-    Route::post('position/repair', 'PositionController@repair');
+    Route::get('position/sanity-checker', 'PositionSanityCheckController@sanityChecker');
+    Route::post('position/repair', 'PositionSanityCheckController@repair');
 
     Route::resource('position', 'PositionController');
